+++ plugins/postcss-custom-media/index.js
@@ -1,29 +1,29 @@
 import postcss from 'postcss';
-import getCustomMedia from './lib/custom-media-from-root';
+import getCustomMediaFromRoot from './lib/custom-media-from-root';
+import getCustomMediaFromImports from './lib/get-custom-media-from-imports';
 import transformAtrules from './lib/transform-atrules';
-import importCustomMediaFromSources from './lib/import-from';
-import exportCustomMediaToDestinations from './lib/export-to';
+import writeCustomMediaToExports from './lib/write-custom-media-to-exports';
 
 export default postcss.plugin('postcss-custom-media', opts => {
-	// whether to preserve custom selectors and rules using them
-	const preserve = Boolean(Object(opts).preserve);
+	// whether to preserve custom media and at-rules using them
+	const preserve = 'preserve' in Object(opts) ? Boolean(opts.preserve) : false;
 
-	// sources to import custom selectors from
+	// sources to import custom media from
 	const importFrom = [].concat(Object(opts).importFrom || []);
 
-	// destinations to export custom selectors to
+	// destinations to export custom media to
 	const exportTo = [].concat(Object(opts).exportTo || []);
 
-	// promise any custom selectors are imported
-	const customMediaPromise = importCustomMediaFromSources(importFrom);
+	// promise any custom media are imported
+	const customMediaPromise = getCustomMediaFromImports(importFrom);
 
 	return async root => {
 		const customMedia = Object.assign(
 			await customMediaPromise,
-			getCustomMedia(root, { preserve })
+			getCustomMediaFromRoot(root, { preserve })
 		);
 
-		await exportCustomMediaToDestinations(customMedia, exportTo);
+		await writeCustomMediaToExports(customMedia, exportTo);
 
 		transformAtrules(root, customMedia, { preserve });
 	};